PR 2.5: spatial collaborator migration
Move the entire Spatial class (~962 LoC) from
src/pyramids/dataset/ops/spatial.py into
src/pyramids/dataset/_collaborators.py:Spatial. Migrated members:

* Public methods: crop, to_crs, set_crs, convert_longitude, resample,
  align, fill_gaps.
* Private methods: _get_crs, _get_epsg, _check_alignment,
  _correct_wrap_cutline_error (staticmethod), _crop_aligned,
  _crop_with_polygon_warp, _crop_with_raster,
  _reproject_with_ReprojectImage.

Substituted self.<state> -> self._ds.<state> for: _check_no_data_value,
_epsg, band_count, cell_size, columns, crs, driver_type, epsg,
gdal_dtype, geotransform, lon, no_data_value, raster, read_array, rows,
top_left_corner.

Six call-site fixups required by the migration:

* Every type(self) constructor / classmethod-access site (10 occurrences)
  becomes self._ds.__class__.<X>(...). The trailing word-boundary in
  the original substitution regex was wrong (')' followed by '.' isn't
  a word boundary), so they had to be rewritten in a follow-up pass.
* _correct_wrap_cutline_error is a staticmethod; calls that previously
  routed through type(self) / base_cls now call Spatial.<X> directly.
* _reproject_with_ReprojectImage assigned reprojected_raster_b = self
  expecting a Dataset; now uses self._ds.
* crop's mask alignment recurses with mask.align(self) which would pass
  a Spatial collaborator; uses self._ds instead.
* fill_gaps's isinstance(self, AbstractDataset) gate now checks
  self._ds (the proxy class).
* Analysis.overlay's self._ds._check_alignment(...) routes through
  self._ds.spatial._check_alignment(...) since the helper migrated.

src/pyramids/dataset/dataset.py:

* Drop _SpatialMixin from import alias list and from Dataset's MRO.
* Add seven facades (crop, to_crs, set_crs, convert_longitude, resample,
  align, fill_gaps) thin-forwarding to self.spatial.<X>.
* AbstractDataset.__init__ calls self._get_epsg() before
  Dataset.__init__ has wired the spatial collaborator, so _get_crs and
  _get_epsg are concrete overrides on Dataset (not facades) — same
  one-liner bodies as on the collaborator. Imports epsg_from_wkt for
  this.

src/pyramids/dataset/ops/__init__.py: drop Spatial. spatial.py: keep
top-level imports only; the Spatial class is gone.

tests/:

* test_dataset_unit.py / test_dataset.py: tests calling private
  Spatial helpers directly (_crop_aligned, _crop_with_polygon_warp,
  _crop_with_raster, _check_alignment, _reproject_with_ReprojectImage,
  _get_crs) now route through ds.spatial.<X>.
  Dataset._correct_wrap_cutline_error -> Spatial._correct_wrap_cutline_error.
* test_collaborators.py: move 7 spatial rows from FORWARDING_METHODS
  to FACADE_METHODS.

Full main test suite passes; only the two pre-existing Windows
file-lock flakes in test_dataset_collection.py remain failing.
