+++ src/pyramids/dataset/ops/__init__.py
@@ -2,10 +2,8 @@
 
 from pyramids.dataset.ops.band_metadata import BandMetadata
 from pyramids.dataset.ops.io import IO
-from pyramids.dataset.ops.spatial import Spatial
 
 __all__ = [
     "BandMetadata",
     "IO",
-    "Spatial",
 ]
